+++ application/modules/mod/config/routes.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * BlizzCMS
+ *
+ * @author  WoW-CMS
+ * @copyright  Copyright (c) 2017 - 2020, WoW-CMS.
+ * @license https://opensource.org/licenses/MIT MIT License
+ * @link    https://wow-cms.com
+ */
+defined('BASEPATH') OR exit('No direct script access allowed');
+
+$route['mod'] = 'mod/index';
+$route['mod/queue'] = 'mod/queue';
+$route['mod/reports'] = 'mod/reports';
+$route['mod/logs'] = 'mod/logs';
+$route['mod/bannings'] = 'mod/bannings';
+$route['mod/warnings'] = 'mod/warnings';